+++ lib/fabrication.rb
@@ -1,3 +1,10 @@
+autoload :Fabricate, 'fabricate'
+
+if defined?(Rake)
+  require 'rake'
+  Dir[File.join(File.dirname(__FILE__), 'tasks', '**/*.rake')].each { |rake| load rake }
+end
+
 module Fabrication
   autoload :DuplicateFabricatorError, 'fabrication/errors/duplicate_fabricator_error'
   autoload :UnfabricatableError,      'fabrication/errors/unfabricatable_error'
@@ -13,7 +20,6 @@ module Fabrication
   end
 
   autoload :Config,     'fabrication/config'
-  autoload :Fabricator, 'fabrication/fabricator'
   autoload :Sequencer,  'fabrication/sequencer'
   autoload :Support,    'fabrication/support'
   autoload :Transform,  'fabrication/transform'
@@ -30,7 +36,7 @@ module Fabrication
   end
 
   def self.clear_definitions
-    @manager = nil
+    manager.clear
     Sequencer.sequences.clear
   end
 
@@ -39,7 +45,7 @@ module Fabrication
   end
 
   def self.manager
-    @manager ||= Fabrication::Schematic::Manager.new
+    @manager ||= Fabrication::Schematic::Manager.instance
   end
 
   def self.schematics
@@ -53,36 +59,8 @@ def Fabricator(name, options={}, &block)
 end
 
 def Fabricate(name, overrides={}, &block)
-  Fabrication::Fabricator.fabricate(name, overrides, &block).tap do |object|
+  Fabricate.create(name, overrides, &block).tap do |object|
     Fabrication::Cucumber::Fabrications[name] = object if Fabrication::Config.register_with_steps?
-  end
-end
-
-class Fabricate
-  def self.times(count, name, overrides={}, &block)
-    count.times.map { Fabricate(name, overrides, &block) }
-  end
-
-  def self.build_times(count, name, overrides={}, &block)
-    count.times.map { Fabricate.build(name, overrides, &block) }
-  end
-
-  def self.attributes_for(name, overrides={}, &block)
-    Fabrication::Fabricator.to_attributes(name, overrides, &block)
-  end
-
-  def self.to_params(name, overrides={}, &block)
-    Fabrication::Fabricator.to_params(name, overrides, &block)
-  end
-
-  def self.build(name, overrides={}, &block)
-    Fabrication::Fabricator.build(name, overrides, &block).tap do |object|
-      Fabrication::Cucumber::Fabrications[name] = object if Fabrication::Config.register_with_steps?
-    end
-  end
-
-  def self.sequence(name=Fabrication::Sequencer::DEFAULT, start=nil, &block)
-    Fabrication::Sequencer.sequence(name, start, &block)
   end
 end