+++ lib/fabrication/support.rb
@@ -7,17 +7,31 @@ class Fabrication::Support
     end
 
     def class_for(class_or_to_s)
-      if class_or_to_s.respond_to?(:to_sym)
-        class_name = variable_name_to_class_name(class_or_to_s)
-        class_name.split('::').inject(Object) do |object, string|
-          object.const_get(string)
+      class_name = variable_name_to_class_name(class_or_to_s)
+      klass = constantize(class_name)
+    rescue NameError => original_error
+      raise Fabrication::UnfabricatableError.new(class_or_to_s, original_error)
+    end
+
+    def constantize(camel_cased_word)
+      names = camel_cased_word.split('::')
+      Object.const_get(camel_cased_word) if names.empty?
+      names.shift if names.size > 1 && names.first.empty?
+      names.inject(Object) do |constant, name|
+        if constant == Object
+          constant.const_get(name)
+        else
+          candidate = constant.const_get(name)
+          next candidate if constant.const_defined?(name, false)
+          next candidate unless Object.const_defined?(name)
+          constant = constant.ancestors.inject do |const, ancestor|
+            break const    if ancestor == Object
+            break ancestor if ancestor.const_defined?(name, false)
+            const
+          end
+          constant.const_get(name, false)
         end
-      else
-        class_or_to_s
       end
-    rescue NameError => e
-      puts e.message
-      puts e.backtrace.join("\n")
     end
 
     def extract_options!(args)
@@ -29,22 +43,18 @@ class Fabrication::Support
     end
 
     def find_definitions
-      Fabrication.manager.preinitialize
-      Fabrication::Config.fabricator_path.each do |folder|
-        Dir.glob(File.join([Fabrication::Config.path_prefix, folder, '**', '*.rb'].compact)).sort.each do |file|
-          load file
-        end
-      end
-    rescue Exception => e
-      puts e.message
-      puts e.backtrace.join("\n")
-      raise e
-    ensure
-      Fabrication.manager.freeze
+      puts "DEPRECATION WARNING: Fabrication::Support.find_definitions has been replaced by Fabrication.manager.load_definitions and will be removed in 3.0.0."
+      Fabrication.manager.load_definitions
     end
 
     def hash_class
       @hash_class ||= defined?(HashWithIndifferentAccess) ? HashWithIndifferentAccess : Hash
+    end
+
+    def singularize(string)
+      string.singularize
+    rescue
+      string.end_with?('s') ? string[0..-2] : string
     end
 
   end