+++ spec/fabrication/generator/active_record_spec.rb
@@ -1,20 +1,22 @@
 require 'spec_helper'
 
-describe Fabrication::Generator::ActiveRecord do
+describe Fabrication::Generator::ActiveRecord, depends_on: :active_record do
 
   describe ".supports?" do
     subject { Fabrication::Generator::ActiveRecord }
+
     it "returns true for active record objects" do
-      subject.supports?(Company).should be_true
+      subject.supports?(ParentActiveRecordModel).should be_true
     end
+
     it "returns false for non-active record objects" do
-      subject.supports?(Person).should be_false
+      subject.supports?(ParentRubyObject).should be_false
     end
   end
 
   describe "#persist" do
     let(:instance) { double }
-    let(:generator) { Fabrication::Generator::ActiveRecord.new(Company) }
+    let(:generator) { Fabrication::Generator::ActiveRecord.new(ParentActiveRecordModel) }
 
     before { generator.send(:_instance=, instance) }
 
@@ -27,35 +29,35 @@ describe Fabrication::Generator::ActiveRecord do
   describe "#create" do
 
     let(:attributes) do
-      Fabrication::Schematic::Definition.new(Company) do
-        name 'Company Name'
-        city { |attrs| attrs[:name].downcase }
-        divisions(count: 2) { |attrs, i| Division.new(name: "Division #{i}") }
+      Fabrication::Schematic::Definition.new(ParentActiveRecordModel) do
+        string_field 'Different Content'
+        number_field { |attrs| attrs[:string_field].length }
+        child_active_record_models(count: 2) { |attrs, i| ChildActiveRecordModel.new(number_field: i) }
       end.attributes
     end
 
-    let(:generator) { Fabrication::Generator::ActiveRecord.new(Company) }
-    let!(:company) { generator.create(attributes, {}) }
-    let(:divisions) { company.divisions }
+    let(:generator) { Fabrication::Generator::ActiveRecord.new(ParentActiveRecordModel) }
+    let!(:parent_active_record_model) { generator.create(attributes, {}) }
+    let(:child_active_record_models) { parent_active_record_model.child_active_record_models }
 
     it 'passes the object to blocks' do
-      company.city.should == 'company name'
+      expect(parent_active_record_model.number_field).to eq 17
     end
 
     it 'passes the object and count to blocks' do
-      company.divisions.map(&:name).should == ["Division 1", "Division 2"]
+      expect(child_active_record_models.map(&:number_field)).to eq [1, 2]
     end
 
     it 'persists the company upon creation' do
-      Company.find_by_name('Company Name').should be
+      expect(ParentActiveRecordModel.where(string_field: 'Different Content').count).to eq 1
     end
 
     it 'generates the divisions' do
-      divisions.length.should == 2
+      expect(child_active_record_models.count).to eq 2
     end
 
     it 'persists the divisions' do
-      Division.find_all_by_company_id(company.id).count.should == 2
+      expect(ChildActiveRecordModel.where(parent_active_record_model_id: parent_active_record_model.id).count).to eq 2
     end
 
   end