+++ spec/fabrication/support_spec.rb
@@ -23,38 +23,33 @@ describe Fabrication::Support do
     context "with a class that doesn't exist" do
 
       it "returns nil for a class name string" do
-        Fabrication::Support.class_for('your_mom').should be_nil
+        lambda { Fabrication::Support.class_for('your_mom') }.
+          should raise_error(Fabrication::UnfabricatableError)
       end
 
       it "returns nil for a class name symbol" do
-        Fabrication::Support.class_for(:your_mom).should be_nil
+        lambda { Fabrication::Support.class_for(:your_mom) }.
+          should raise_error(Fabrication::UnfabricatableError)
       end
 
-    end
+      context "and custom const_missing is defined" do
+        before do
+          module Family
+            def self.const_missing(name)
+              raise NameError, "original message"
+            end
+          end
+        end
 
+        it "raises an exception with the message from the original exception" do
+          lambda { Fabrication::Support.class_for("Family::Mom") }.
+            should raise_error(Fabrication::UnfabricatableError, /original message/)
+        end
+      end
+    end
   end
 
-  describe ".find_definitions" do
-    before { Fabrication.clear_definitions }
-
-    context 'happy path' do
-      it "loaded definitions" do
-        Fabrication::Support.find_definitions
-        Fabrication.manager[:parent_ruby_object].should be
-      end
-    end
-
-    context 'when an error occurs during the load' do
-      it 'still freezes the manager' do
-        Fabrication::Config.should_receive(:fabricator_path).and_raise(Exception)
-        expect { Fabrication::Support.find_definitions }.to raise_error
-        Fabrication.manager.should_not be_initializing
-      end
-    end
-
-  end
-
-  describe '.hash_class' do
+  describe '.hash_class', depends_on: :active_support do
     subject { Fabrication::Support.hash_class }
 
     context 'with HashWithIndifferentAccess defined' do