+++ spec/fabrication/syntax/make_spec.rb
@@ -3,48 +3,46 @@ require 'fabrication/syntax/make'
 
 describe Fabrication::Syntax::Make do
 
-  describe "#make mongoid" do
-
+  describe "#make mongoid", depends_on: :mongoid do
     it "should return a fabricated object" do
-      Author.make.should be_instance_of Author
+      ParentMongoidDocument.make.should be_instance_of ParentMongoidDocument
     end
 
     it "should overwrite options" do
-      Author.make(:name => "N.Rodrigues").name.should eql("N.Rodrigues")
+      ParentMongoidDocument.make(string_field: "N.Rodrigues").string_field.should eql("N.Rodrigues")
     end
 
     it "should treat a first non-hash argument as fabrication name suffix" do
-      Author.make(:with_handle, :name => "Eric Arthur Blair").handle.should eql("@1984")
+      Fabricator(:parent_mongoid_document_with_handle, from: :parent_mongoid_document)
+      ParentMongoidDocument.make(:with_handle).string_field.should eql("content")
     end
 
     it "should work the same as Fabricate.build" do
-      Author.make.should be_new_record
+      ParentMongoidDocument.make.should be_new_record
     end
 
     it "bang should be the same as Fabricate" do
-      Author.make!.should_not be_new_record
+      ParentMongoidDocument.make!.should_not be_new_record
     end
-
   end
 
-  describe "#make activerecord" do
+  describe "#make activerecord", depends_on: :active_record do
 
     it "should return a fabricated object" do
-      Company.make.should be_instance_of Company
+      ParentActiveRecordModel.make.should be_instance_of ParentActiveRecordModel
     end
 
     it "should work the same as Fabricate.build" do
-      Company.make.should be_new_record
+      ParentActiveRecordModel.make.should be_new_record
     end
 
     it "bang should be the same as Fabricate" do
-      Company.make!.should_not be_new_record
+      ParentActiveRecordModel.make!.should_not be_new_record
     end
 
   end
 
-  describe "#make sequel" do
-
+  describe "#make sequel", depends_on: :sequel do
     it "should return a fabricated object" do
       ParentSequelModel.make.should be_instance_of ParentSequelModel
     end