+++ spec/support/plain_old_ruby_objects.rb
@@ -48,44 +48,15 @@ class ChildRubyObject < Persistable
   end
 end
 
+module NamespacedClasses
+  class RubyObject < OpenStruct
+    attr_accessor :display
+  end
+end
+
 class Troublemaker
   def raise_exception=(value)
     raise "Troublemaker exception" if value
-  end
-end
-
-class Dog
-  attr_accessor :name, :breed, :locations
-end
-
-class Location
-  attr_accessor :lat, :lng
-end
-
-class Person
-  attr_accessor :age, :first_name, :last_name, :shoes, :location
-end
-
-class City
-  attr_accessor :city, :state
-
-  def initialize(city, state)
-    self.city = city
-    self.state = state
-  end
-end
-
-class Address
-  attr_accessor :city, :state
-end
-
-class Contact
-  attr_accessor :name, :address
-end
-
-module Something
-  class Amazing
-    attr_accessor :stuff
   end
 end